BUGFIX
- Readme badge edit
- docker-compose example image
- remove config delete in container

+++ app.js
@@ -170,7 +170,9 @@ async function checkLogin(page) {
     }
     console.log('\n🛑 Login failed!');
     console.log('🔑 Wrong token!');
-    fs.unlinkSync(configPath);
+    if (!process.env.token){
+      fs.unlinkSync(configPath);
+    }
     process.exit()
 }
